Suspend monitoring of seek, pause and resume
As with our local copy of v1.4 of this code, we do not wish to trigger
events for when users seek, pause or resume playback.

For seek and pause, we can avoid listening for those JS events
altogether. For resume, which uses thesame event listener as playback,
we suspend the tiggering of the event sending.